+++ CHANGES.md
@@ -3,6 +3,7 @@ This files lists the changes during the lifetime of this project.
 
 ## unreleased
 - (backwards incompatible) clean slate - start over with only essential 4W functionality
+- (backwards incompatible) merge User and Colleague model
 
 ## 2025-10-09
 - add period filter to placement page